Removed old dead code, tried to solve a bug, but surmise that VC2008 /GL is culprit.

+++ src/optionaff.h
@@ -46,8 +46,8 @@ struct optionStruct
     bool computeParms;// compute parms
     bool suffixOnly;// suffix only
     bool verbose;         // verbose
-    int minperc;
-    int maxperc;
+    double minperc;
+    double maxperc;
     bool doweights;
     optionStruct();
     ~optionStruct();
